[FLINK-19093][task] Fix isRunning check in AsyncCheckpointRunnable registration
Currently, SubtaskCheckpointCoordinator#close call will close all the AsyncCheckpointRunnables outside of the synchronized block. When the respective AsyncCheckpointRunnable is registering meanwhile, it would throw an exception since the coordinator was already closed but the internal runnable is still running.

After this commit, when runnable registration with closed coordinator, the check of isRunning state is replaced by checking whether the respective runnable still exists in the checkpoint map or not, to avoid unexpected exception thrown.

This closes #13286.