[FLINK-21610][tests] Harden ProcessFailureCancelingITCase
The cancellation of the job may fail if the job termination finishes so quickly that the job has been cleaned up before the cancellation has been processed.
The cancellation is unnecessary anyway because the TM failure causes the job to fail and we explicitly forbid restarts.
As such we can just remove the cancel call.

Furthermore, the conditions for when to kill the TM were not correct, as they did not ensure the job was actually deployed.