[FLINK-22864][table] Remove the flink-table-planner and flink-table-uber module
In order to avoid user confusion, reduce duplicate code, and improve maintainability
and testing times of the Flink project as a whole this removes the legacy planner
module.

It means that both the BatchTableEnvironment and DataSet API interop with
Table API are reaching end of life. Use the unified TableEnvironment for
batch and stream processing with the new planner or the DataStream API
in batch execution mode.

Users are encouraged to update their pipelines. Flink 1.13 is the last
version that offers the old functionality.

This closes #16080.