[FLINK-23811][tests] Handle finished subtasks in CommonTestUtils.waitForAllTaskRunning
CommonTestUtils.waitForAllTaskRunning returns when all the subtasks are running AND
the job is running and not finished. However, with FLIP-147, subtasks may finish and
the job will still be running. So the method won't return and instead timeout.

This commit adds a flag to indicate whether to fail or proceed
when a finished sub-task is encountered.